server/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
click==8.1.7
colorama==0.4.6
fastapi==0.109.0
h11==0.14.0
idna==3.6
numpy==1.26.3
opencv-python==4.9.0.80
pydantic==2.6.0
pydantic_core==2.16.1
sniffio==1.3.0
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.27.0.post1
httpx
pillow
python-multipart
python-dotenv
supabase
pyjwt
orjson